+++ tests/testthat/test_G20.R
@@ -1,5 +1,16 @@
-context("G20")
+context("G20 data frame")
 
-test_that("G20 dataset is included", {
-  expect_equal(names(G20), c("region", "country", "gdp_mil_usd", "hdi", "econ_classification", "hemisphere"))
+test_that("G20 data frame is present", {
+  expect_equal(
+    names(G20),
+    c(
+      "region",
+      "country",
+      "gdp_mil_usd",
+      "hdi",
+      "econ_classification",
+      "hemisphere"
+    )
+  )
+  expect_equal(nrow(G20), 20)
 })